backend/requirements.txt
beautifulsoup4>=4.10.0,<=4.12.3
charset-normalizer>=2.0.12,<=3.2.0
Cython>=0.29.28,<=3.0.2
dedoc-utils==0.3.8
fastapi>=0.77.0,<1.0
huggingface-hub>=0.14.1,<1.0
imutils==0.5.4
itsdangerous>=2.1.0,<=2.1.2
numpy>=1.22.0,<2.0
olefile~=0.46
opencv-python>=4.5.5.64,<4.6.0
orjson>=3.8.11,<4.0
pandas>=1.4.1,<3
pdf.tocgen>=1.3.0,<=1.3.4
pdf2image==1.10.0 #1.14.0 - there are converting artifacts '№' != '№\n\x0c'
pdfminer.six>=20211012,<=20231228
piexif==1.1.3
puremagic>=1.0,<2.0 # needs libmagic to be installed in the system
pylzma==0.5.0
pypdf>=4.0,<6.0
pytesseract==0.3.10
python-docx==0.8.11
python-Levenshtein==0.12.2
python-logstash-async>=2.5.0,<=2.7.0
python-magic<1.0
python-multipart==0.0.6
rarfile==4.0
requests>=2.22.0
roman>=3.3,<4.0
scikit-image>=0.19.3,<=0.21.0
scikit_learn>=1.0.2,<=1.2.2
scipy>=1.8.0,<=1.11.2
six>=1.14.0,<2.0
starlette>=0.26.1,<=0.27.0
texttable>=1.6.7,<2.0
ujson>=5.4.0,<=5.8.0
uvicorn>=0.18.0,<=0.23.2
wget==3.2
xgbfir>=0.3.1,<1.0
xgboost>=1.6.0,<2.0  # lower versions aren't compatible with pandas>2
xlrd>=1.2.0,<2.0
nltk==3.9.1
fontTools==4.56.0
torch~=1.11.0
torchvision~=0.12.0
uvicorn
fastapi
python-multipart
